+++ src/main/java/net/geforcemods/securitycraft/mixin/camera/ServerPlayerEntityMixin.java
@@ -1,20 +1,38 @@
 package net.geforcemods.securitycraft.mixin.camera;
 
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
+import net.geforcemods.securitycraft.entity.camera.SecurityCamera;
 import net.geforcemods.securitycraft.util.PlayerUtils;
+import net.minecraft.entity.Entity;
 import net.minecraft.entity.player.ServerPlayerEntity;
 
-/**
- * Makes sure the server does not move the player mounting a camera to the camera's position
- */
 @Mixin(value = ServerPlayerEntity.class, priority = 1100)
-public class ServerPlayerEntityMixin {
+public abstract class ServerPlayerEntityMixin {
+	@Shadow
+	public abstract Entity getCamera();
+
+	/**
+	 * Makes sure the server does not move the player mounting a camera to the camera's position
+	 */
 	@Redirect(method = "tick", at = @At(value = "INVOKE", target = "Lnet/minecraft/entity/player/ServerPlayerEntity;absMoveTo(DDDFF)V"))
 	private void securitycraft$tick(ServerPlayerEntity player, double x, double y, double z, float yaw, float pitch) {
 		if (!PlayerUtils.isPlayerMountedOnCamera(player))
 			player.absMoveTo(x, y, z, yaw, pitch);
+	}
+
+	/**
+	 * Ensures that players in spectator mode are able to see players that are currently viewing a camera, because players
+	 * that are spectating another entity are usually invisible for other spectators
+	 */
+	@Inject(method = "broadcastToPlayer", at = @At("HEAD"), cancellable = true)
+	private void securitycraft$broadcastPlayerViewingCamera(ServerPlayerEntity player, CallbackInfoReturnable<Boolean> cir) {
+		if (getCamera() instanceof SecurityCamera && player.isSpectator())
+			cir.setReturnValue(true);
 	}
 }